[-] BO : #PSCFI-2902 - Fixed wrong URL [-] BO : #PSCFI-2922 - Fixed bugs in GroupReduction. Now, group_reduction and product_group_reduction tables are now populated/cleaned properly (CRUD Products/Category).

+++ classes/Product.php
@@ -315,7 +315,7 @@ class ProductCore extends ObjectModel
 		if ($this->id_category_default)
 			$this->category = Category::getLinkRewrite((int)$this->id_category_default, (int)$id_lang);
 	}
-
+	
 	public function getFields()
 	{
 		parent::validateFields();
@@ -402,7 +402,7 @@ class ProductCore extends ObjectModel
 		}
 		return $fields;
 	}
-
+	
 	public static function initPricesComputation($id_customer = NULL)
 	{
 		global $cookie;
@@ -549,6 +549,9 @@ class ProductCore extends ObjectModel
 
 	public function delete()
 	{
+		if (!GroupReduction::deleteProductReduction($this->id))
+			return false;
+		
 		Hook::deleteProduct($this);
 		if (!parent::delete() OR
 			!$this->deleteCategories(true) OR
@@ -679,6 +682,9 @@ class ProductCore extends ObjectModel
 		if (!$this->addToCategories($categories))
 			return false;
 
+		if (!$this->setGroupReduction($categories))
+			return false;
+		
 		return true;
 	}
 
@@ -3365,6 +3371,28 @@ class ProductCore extends ObjectModel
 		UPDATE `'._DB_PREFIX_.'product`
 		SET ecotax = 0
 		');
+	}
+	
+	/**
+	 * Set Group reduction if needed
+	 * @param string $productCategories Categories the product belongs to
+	 */
+	public function setGroupReduction($categories)
+	{
+		if (sizeof($categories) > 1) // Use category_default id if more than 1 category
+			$row = GroupReduction::getGroupByCategoryId((int)$this->id_category_default);
+		else
+			$row = GroupReduction::getGroupByCategoryId((int)array_pop($categories));
+		
+		if (!$row) { // Then Remove
+			if (!GroupReduction::deleteProductReduction($this->id))
+				return false;
+		}
+		else {
+			if (!GroupReduction::setProductReduction($this->id, $row['id_group'], $category, $row['reduction'])) // Add or Edit
+				return false;
+		}
+		return true;
 	}
 }
